Python Full Stack - Internship (May - Jul 2025)/SpellChecker (Desktop App)/requirements.txt
pyqt5 
pyspellchecker 
language-tool-python 
python-docx